Update latest changes

+++ nbbuild/antsrc/org/netbeans/nbbuild/CustomJavac.java
@@ -20,9 +20,12 @@
 package org.netbeans.nbbuild;
 
 import java.io.File;
+import java.net.MalformedURLException;
+import java.net.URISyntaxException;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
 import org.apache.tools.ant.BuildException;
@@ -67,12 +70,16 @@ public class CustomJavac extends Javac {
     @Override
     public void execute() throws BuildException {
         String release = getRelease();
-        if (release == null || release.isEmpty()) {
+        if (release != null && release.isEmpty()) {
+            setRelease(null); // unset property
+            release = null;
+        }
+        if (release == null) {
             String tgr = getTarget();
             if (tgr.matches("\\d+")) {
                 tgr = "1." + tgr;
             }
-            if (!isBootclasspathOptionUsed()) {
+            if (canUseRelease()) {
                 setRelease(tgr.substring(2));
             }
             String src = getSource();
@@ -115,16 +122,28 @@ public class CustomJavac extends Javac {
         } else {
             log("Warning: could not create " + generatedClassesDir, Project.MSG_WARN);
         }
+        try {
+            Class<?> mainClazz = CustomJavacClassLoader.findMainCompilerClass(getProject());
+            if (mainClazz != null) {
+                super.add(CustomJavacClassLoader.createCompiler(mainClazz, canUseRelease()));
+            }
+        } catch (ClassNotFoundException | MalformedURLException | URISyntaxException ex) {
+            if (ex instanceof BuildException) {
+                throw (BuildException) ex;
+            }
+            throw new BuildException(ex);
+        }
         super.compile();
     }
 
-    private boolean isBootclasspathOptionUsed() {
+    private boolean canUseRelease() {
+        // 'error: exporting a package from system module java.desktop is not allowed with --release'
         for (String arg : getCurrentCompilerArgs()) {
-            if (arg.contains("-Xbootclasspath")) {
-                return true;
+            if (arg.contains("-Xbootclasspath") || arg.contains("--add-exports=java.")) {
+                return false;
             }
         }
-        return false;
+        return true;
     }
 
     /**
@@ -226,12 +245,16 @@ public class CustomJavac extends Javac {
                 continue;
             }
             int i = clazz.indexOf('$');
-            File enclosing = new File(d, clazz.substring(0, i) + ".class");
-            if (!enclosing.isFile()) {
-                File enclosed = new File(d, clazz);
-                log(clazz + " will be deleted since " + enclosing.getName() + " is missing", Project.MSG_VERBOSE);
-                if (!enclosed.delete()) {
-                    throw new BuildException("could not delete " + enclosed, getLocation());
+            // ignore filenames that start right with '$' (separatorChar preceded), these could not be inner classes.
+            if (i > 0 && clazz.charAt(i - 1) != File.separatorChar) {
+                File enclosing = new File(d, clazz.substring(0, i) + ".class");
+                // no inner class' filename may begin directly with '$', it must be preceded by an outer class' name.
+                if (!enclosing.isFile()) {
+                    File enclosed = new File(d, clazz);
+                    log(clazz + " will be deleted since " + enclosing.getName() + " is missing", Project.MSG_VERBOSE);
+                    if (!enclosed.delete()) {
+                        throw new BuildException("could not delete " + enclosed, getLocation());
+                    }
                 }
             }
         }
@@ -247,5 +270,4 @@ public class CustomJavac extends Javac {
         }
         return false;
     }
-
 }
